Make .editorconfig path relative so editorconfig module can find it
Before this change, tests/core/editorconfig.js was failing, because
the editorconfig module was concatenating the folder path with the
absolute .editorconfig path, causing an invalid path.

This change makes the .editorconfig "path" relative, so the editorconfig
plugin simply traverses up the directory tree until it finds it in
the root of the project.